Updated for timeout usage mode
Originally it would result in a double callback execution. Depending on use case, that could be negligible. But it was very important in my case. Thus this pull request/commit.
Adjustment makes use of contextual call()'s to prevent any potential for undefined references. 
And now utilizes a wasSuccess logical check, to prevent double callback. 

The idea is that if the "event" (aka pingCheck) is a result of a timeout, then wasSuccess is the default value of false.
And in "timeout mode", the pingCheck is not called (since the timeout calls it anyway!), in which case the wasSuccess would be marked true prior to execution of timeout followup call.

The only issue I'm aware of with this, is that if timeout mode is on, and the timeout value is lower than the network response, then it will fail as though the ping check couldn't complete, but in actuality it succeeded, just too slowly.
I haven't needed a fix for that (ie. just increase the timeout value based on expected network speed), so therefore am leaving that issue unapproached.